% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.R
\name{draws}
\alias{draws}
\title{Extract subset of posterior draws}
\usage{
draws(fit, pars, as_draws_df = FALSE, parallel = TRUE)
}
\arguments{
\item{fit}{Fitted Stan model (a \code{CmdStanMCMC} object.)}

\item{pars}{Character vector of regular expressions capturing the parameters
of interest. See Details below.}

\item{as_draws_df}{Whether to convert the result to \link[posterior:draws_df]{posterior::draws_df} format.
Requires that the \code{posterior} package is installed.}

\item{parallel}{Whether to scan files in parallel using \link[parallel:mclapply]{parallel::mclapply}.}
}
\value{
Data frame with one row per draw and one column per
parameter. If \code{as_draws_df} is TRUE, this is then passed to
\link[posterior:draws_df]{posterior::as_draws_df}, yielding a draws object compatible with
\code{posterior} and related packages.
}
\description{
Load draws of a subset of parameters from filtered CmdStan output files.
This tends to be more time and memory efficient than calling \verb{$draws()}
on the associated fit object.
}
\details{
Indexed parameters are renamed in the CSV files, with the opening bracket and
commas replaced by full stops and the closing bracket dropped. For example,
to include draws of "eta[3,4]", \code{pars} needs to capture "eta.3.4" instead.
}
\seealso{
\link{stansummary}.
}
