% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Convert a summary of draws to a matrix}
\usage{
extract_matrix(df, key = "name", value = "Mean", cnames = NULL)
}
\arguments{
\item{df}{Data frame with posterior summaries of matrix entries. Typically
a result of asking for a \link{stansummary} or \link[cmdstanr:fit-method-summary]{cmdstanr::summary} of a matrix
parameter. Should not contain summaries of any other parameters.}

\item{key}{The column in \code{df} with the parameter name and indexes.
The default corresponds to the output format of \link{stansummary}.}

\item{value}{The column in \code{df} from which the values are to be read into
the matrix. The default corresponds to the posterior mean as reported by \link{stansummary}.}

\item{cnames}{Column names to be applied to the resulting matrix. Needs to
be a character vector of the length corresponding to the number of columns
in the matrix. If this condition is not met or if the argument is NULL
(the default), generic X1,...,Xn names are used.}
}
\value{
A matrix.
}
\description{
Arrange a posterior summary of parameters that are entries of a matrix
or of a 2D array into an R matrix object.
}
\seealso{
\code{\link[=process_array_summary]{process_array_summary()}}.
}
