% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fapply}
\alias{fapply}
\title{Optionally parallelised \code{lapply}.}
\usage{
fapply(X, FUN, parallel = FALSE, ...)
}
\arguments{
\item{X}{List or vector of arguments.}

\item{FUN}{Function to apply to these arguments.}

\item{parallel}{If FALSE (the default), call \code{lapply}. Otherwise call
\code{parallel::mclapply}.}

\item{...}{Additional parameters passed to the wrapped function. See Details below.}
}
\value{
The result of calling the wrapped function.
}
\description{
Excecute either \code{lapply} or \code{parallel::mclapply}, depending on the value
of a Boolean flag.
}
\details{
When \code{parallel} is TRUE and \code{...} does not specify \code{mc.cores},
it is set to the number of available cores minus one, or to the
length of \code{X}, whichever is smaller.
}
