% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansummary.R
\name{filter_cmdstan_csv}
\alias{filter_cmdstan_csv}
\title{Discard unwanted CSV columns by streaming}
\usage{
filter_cmdstan_csv(src, pars, dst = tempfile())
}
\arguments{
\item{src}{Path to the CSV file to process.}

\item{pars}{Vector of regular expressions or strings identifying the columns
to keep.}

\item{dst}{Path to the processed CSV file.}
}
\value{
\code{dst}.
}
\description{
Forms and executes system commands resulting in
the \code{src} CSV being converted to \code{dst} by stripping comments and
keeping only the columns specified in \code{pars}.
}
