% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansummary.R
\name{filter_cmdstan_csvs}
\alias{filter_cmdstan_csvs}
\title{Slim down a collection of CSV files}
\usage{
filter_cmdstan_csvs(
  src,
  pars,
  dst = replicate(length(src), tempfile()),
  parallel = TRUE
)
}
\arguments{
\item{src}{Vector of paths to the CSV files to be processed.}

\item{pars}{Vector of regular expressions or strings identifying the columns
to keep.}

\item{dst}{Vector of paths where the processed CSVs are to be written.}

\item{parallel}{Whether to parallelise the operation with \link[parallel:mclapply]{parallel::mclapply}.}
}
\value{
\code{dst}
}
\description{
Apply \code{filter_cmdstan_csv} to a collection of CSV files,
optionally in parallel.
}
\seealso{
\link{filter_cmdstan_csv}
}
