% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_cmdstan_fit}
\alias{inspect_cmdstan_fit}
\title{Quick look at a CmdStan fit}
\usage{
inspect_cmdstan_fit(
  cmdstan_csv_dir,
  file_regex = ".*",
  iter_ignore = 0,
  parallel = TRUE,
  print = TRUE
)
}
\arguments{
\item{cmdstan_csv_dir}{Directory to which CmdStan wrote or is writing the samples.}

\item{file_regex}{Regular expression to filter file names in \code{cmdstan_csv_dir}.
Useful if multiple fits are being stored in the same directory. The default
expression matches any file name.}

\item{iter_ignore}{Number of initial iterations to discard
before calculating summary statistics. Useful for skipping warmup iterations.}

\item{parallel}{Whether to process CmdStan CSV files in parallel with \link[parallel:mclapply]{parallel::mclapply}.}

\item{print}{Whether to print the statistics on screen.}
}
\value{
A data frame of statistics, invisibly.
}
\description{
This function provides basic statistics about a CmdStan fit such as
the count of divergences and the average treedepth. The intended use
is to help the user decide whether to abort a fit that is in progress.
}
\note{
\itemize{
\item The chain number reported may not match the number assigned by CmdStan.
\item The iteration number reported is the number of samples found
in the file minus \code{iter_ignore}. When \code{iter_ignore} is 0 and CmdStan does not
save warmup samples, it is the number of completed sampling iterations.
}
}
