% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesummary.R
\name{mlesummary}
\alias{mlesummary}
\title{Extract estimates from MLE fit}
\usage{
mlesummary(fit, pars = NULL)
}
\arguments{
\item{fit}{CmdStan model fitted by penalised maximum likelihood (a \code{CmdStanMLE} object.)}

\item{pars}{Character vector of regular expressions selecting
the parameters to report. See the Details section of \link{stansummary} for
an important note on how to capture elements of vector and array parameters.}
}
\value{
Data frame with \code{variable} and \code{estimate} columns.
}
\description{
This function reads in parameter estimates from a \code{CmdStanMLE} object,
mirroring the output of the \verb{$summary()} function in \code{cmdstanr}.
}
\seealso{
\code{\link[=stansummary]{stansummary()}}.
}
