% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{process_array_summary}
\alias{process_array_summary}
\title{Extract and map array indices}
\usage{
process_array_summary(df, key = "name", names = NULL, labels = NULL)
}
\arguments{
\item{df}{Data frame with a character column containing variable names
of the form \code{str[int,...,int]}, for example \code{phi[1,10,2]},
typically the output of \link[cmdstanr:fit-method-summary]{cmdstanr::summary} or \link{stansummary}.
A single array summary is expected. Processing multiple arrays at once
will succeed only if all have the same dimension.}

\item{key}{Name of the column with array variable names. The default
corresponds to the output of \link{stansummary}.}

\item{names}{Vector of strings to use as new column names.
Its length must match the array dimension.
When NULL, lower-case letters starting with "i" are used.}

\item{labels}{List of vectors of labels to use instead of array indices.
Length of the list must match array dimension. When NULL, no replacing takes
place. To replace along some dimensions but not others, pass the empty vector \code{c()}
at the correct position. See Examples below.}
}
\value{
Data frame with additional columns appended.
}
\description{
Extract indices of a posterior summary of an array of parameters to separate
columns, optionally mapping them to labels.
}
\examples{
x <- data.frame(
  name = c("phi[1,1]", "phi[1,2]", "phi[2,1]", "phi[2,2]", "phi[3,1]", "phi[3,2]"),
  Mean = c(2830, 2833, 2790, 2791, 2700, 2730)
)
#       name Mean
# 1 phi[1,1] 2830
# 2 phi[1,2] 2833
# 3 phi[2,1] 2790
# 4 phi[2,2] 2791
# 5 phi[3,1] 2700
# 6 phi[3,2] 2730

process_array_summary(x, names = c("player", "time"), labels = list(c("Carlsen", "Caruana", "Gukesh"), c()))
#       name Mean  player time
# 1 phi[1,1] 2830 Carlsen    1
# 2 phi[1,2] 2833 Carlsen    2
# 3 phi[2,1] 2790 Caruana    1
# 4 phi[2,2] 2791 Caruana    2
# 5 phi[3,1] 2700  Gukesh    1
# 6 phi[3,2] 2730  Gukesh    2

}
