% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansummary.R
\name{read_csv_header}
\alias{read_csv_header}
\title{Read CSV column names}
\usage{
read_csv_header(filepath, comment = "#")
}
\arguments{
\item{filepath}{Path to the CSV file.}

\item{comment}{Comment prefix.}
}
\value{
The first non-commented line, presumably the header, as a character vector.
}
\description{
Read a CSV file in line by line, skipping comments until
the header is encountered.
}
