% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repair_dotnames}
\alias{repair_dotnames}
\title{Convert internal CmdStan parameter names.}
\usage{
repair_dotnames(dotnames)
}
\arguments{
\item{dotnames}{Character vector of internal parameter names.}
}
\value{
Vector of matching length with corrected names.
}
\description{
CmdStan replaces the square brackets \code{[} and commas \verb{,} by dots \code{.}
in the CSVs with model parameter estimates or posterior draws. This
function restores the square bracket notation, so that eta.3.4 becomes
eta\link{3,4}.
}
