% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stansummary.R
\name{stansummary}
\alias{stansummary}
\title{Summarise posterior draws efficiently}
\usage{
stansummary(fit, pars = NULL, sig_figs = 3, parallel = TRUE)
}
\arguments{
\item{fit}{A \code{CmdStanMCMC} or \code{CmdStanVB} object.}

\item{pars}{Character vector of regular expressions or strings
selecting the parameters to summarise.
See Details below. When NULL (default), all parameters are included.}

\item{sig_figs}{The number of significant figures to use in the output,
as interpreted by \code{stansummary}.}

\item{parallel}{Whether subsetting of CmdStan CSV files should be done
in parallel using \link[parallel:mclapply]{parallel::mclapply}.}
}
\value{
Data frame with posterior summary statistics for parameter of interest.
}
\description{
Calculate a table of posterior summaries similar to that
provided by \link[cmdstanr:fit-method-summary]{cmdstanr::summary} but using CmdStan's
\code{stansummary} command-line utility instead of the \code{posterior} package.
This is much faster for models with a very large number of parameters.
Additional time and memory savings can be achieved by subsetting parameters.
}
\details{
When called with \code{pars = NULL}, this function runs
\code{cmdstanr::cmdstan_path()/bin/stansummary} on \code{fit$output_files()} via
a \code{\link{system2}} call, writing the output to a temporary CSV file and
reading it back in.

When \code{pars} is not \code{NULL}, the function scans the header of the CmdStan
CSVs for matches using \code{\link{grepl}}. Note that the indexed variables are
renamed in these CSVs so that "eta[3,4]" becomes "eta.3.4", and \code{pars}
has to be set accordingly.
Once the columns of interest are identified, slimmed-down temporary
CSVs with just the parameters of interest are passed to \code{stansummary}.
The pre-processing of CSVs is done by streaming them through \code{grep} and \code{cut},
and can be done in parallel at a small memory cost.

The parameter subsetting functionality requires sufficient disk space to store
temporary files, up to twice as much as taken by the CmdStan CSVs. These
files are cleared as soon as possible and in any event do not persist
beyond the lifetime of the R session that created them.

This code has been tested on Linux only.
}
