% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vgrepl}
\alias{vgrepl}
\title{grepl vectorized over patterns}
\usage{
vgrepl(patterns, x)
}
\arguments{
\item{patterns}{Vector of regular expressions or strings.}

\item{x}{Character vector to scan for matches.}
}
\value{
Boolean vector.
}
\description{
Return a Boolean mask over \code{x} identifying elements that match at least one
of \code{patterns}.
}
\seealso{
\link{grepl}.
}
